Put Seismic HTTP providers behind reqwest feature flag; move seismic layer into module (#52)
