+++ crates/provider/src/layers/seismic/http.rs
@@ -1,21 +1,14 @@
-//! A provider layer that uses for filling sesimic transactions
-
+//! Seismic provider for HTTP requests
 use crate::{
     fillers::{
         FillProvider, JoinFill, NonceFiller, RecommendedFillers, SimpleNonceManager, WalletFiller,
     },
-    Identity, PendingTransactionBuilder, Provider, ProviderBuilder, ProviderLayer, RootProvider,
-    SendableTx,
+    Identity, ProviderBuilder, RootProvider,
 };
-use alloy_consensus::{transaction::TxSeismicElements, TxSeismic};
-use alloy_network::{Ethereum, EthereumWallet, Network, TransactionBuilder};
-use alloy_primitives::{Address, Bytes, TxKind};
-use alloy_rpc_types_eth::{TransactionInput, TransactionRequest};
-use alloy_transport::{Transport, TransportErrorKind, TransportResult};
-use std::{marker::PhantomData, ops::Deref};
+use alloy_network::{Ethereum, EthereumWallet};
+use std::ops::Deref;
 
-#[cfg(feature = "ws")]
-use alloy_transport::TransportError;
+use crate::layers::seismic::{layer::SeismicLayer, provider::SeismicProvider};
 
 /// Seismic provider
 pub type SeismicSignedProviderInner = FillProvider<
@@ -117,182 +110,7 @@ impl Deref for SeismicUnsignedProvider {
     }
 }
 
-/// Seismic unsigned websocket provider
-#[cfg(feature = "ws")]
-pub type SeismicUnsignedWsProviderInner = RootProvider<alloy_transport::BoxTransport>;
-
-#[cfg(feature = "ws")]
-/// Seismic unsigned websocket provider
-#[derive(Debug, Clone)]
-pub struct SeismicUnsignedWsProvider(pub SeismicUnsignedWsProviderInner);
-
-#[cfg(feature = "ws")]
-impl SeismicUnsignedWsProvider {
-    /// creates a new websocket provider for a client
-    pub async fn new(url: impl Into<String>) -> Result<Self, TransportError> {
-        let provider = ProviderBuilder::new().on_builtin(&url.into()).await?;
-        Ok(Self(provider))
-    }
-
-    /// Get the inner provider
-    pub fn inner(&self) -> &SeismicUnsignedWsProviderInner {
-        &self.0
-    }
-}
-
-/// Get a seismic transaction builder
-pub fn build_seismic_tx(plaintext: Bytes, to: TxKind, from: Address) -> TransactionRequest {
-    TransactionRequest {
-        from: Some(from),
-        to: Some(to),
-        input: TransactionInput { input: Some(plaintext), data: None },
-        transaction_type: Some(TxSeismic::TX_TYPE),
-        ..Default::default()
-    }
-}
-
-/// Seismic middlware for encrypting transactions and decrypting responses
-#[derive(Debug, Clone)]
-pub struct SeismicLayer {}
-
-impl<P, T, N> ProviderLayer<P, T, N> for SeismicLayer
-where
-    P: Provider<T, N>,
-    T: Transport + Clone,
-    N: Network,
-{
-    type Provider = SeismicProvider<P, T, N>;
-
-    fn layer(&self, inner: P) -> Self::Provider {
-        SeismicProvider::new(inner)
-    }
-}
-
-/// Seismic middlware for encrypting transactions and decrypting responses
-#[derive(Debug, Clone)]
-pub struct SeismicProvider<P, T, N> {
-    /// Inner provider.
-    inner: P,
-    /// Phantom data
-    _pd: PhantomData<(T, N)>,
-}
-
-impl<P, T, N> SeismicProvider<P, T, N>
-where
-    P: Provider<T, N>,
-    T: Transport + Clone,
-    N: Network,
-{
-    /// Create a new seismic provider
-    fn new(inner: P) -> Self {
-        Self { inner, _pd: PhantomData }
-    }
-
-    /// Should encrypt input
-    pub fn should_encrypt_input<B: TransactionBuilder<N>>(&self, tx: &B) -> bool {
-        tx.input().map_or(false, |input| !input.is_empty()) && tx.nonce().is_some()
-    }
-}
-
-/// Implement the Provider trait for the SeismicProvider
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
-impl<P, T, N> Provider<T, N> for SeismicProvider<P, T, N>
-where
-    P: Provider<T, N>,
-    T: Transport + Clone,
-    N: Network,
-{
-    fn root(&self) -> &RootProvider<T, N> {
-        self.inner.root()
-    }
-
-    async fn seismic_call(&self, mut tx: SendableTx<N>) -> TransportResult<Bytes> {
-        if let Some(builder) = tx.as_mut_builder() {
-            if self.should_encrypt_input(builder) {
-                let network_pk = self.inner.get_tee_pubkey().await.map_err(|e| {
-                    TransportErrorKind::custom_str(&format!(
-                        "Error getting tee pubkey from server: {:?}",
-                        e
-                    ))
-                })?;
-                let encryption_keypair = TxSeismicElements::get_rand_encryption_keypair();
-                let seismic_elements = TxSeismicElements::default()
-                    .with_encryption_pubkey(encryption_keypair.public_key())
-                    .with_encryption_nonce(TxSeismicElements::get_rand_encryption_nonce());
-
-                // Encrypt using recipient's public key and generated private key
-                let plaintext_input = builder.input().unwrap();
-                let encrypted_input = seismic_elements
-                    .client_encrypt(&plaintext_input, &network_pk, &encryption_keypair.secret_key())
-                    .map_err(|e| {
-                        TransportErrorKind::custom_str(&format!("Error encrypting input: {:?}", e))
-                    })?;
-
-                builder.set_input(Bytes::from(encrypted_input));
-                builder.set_seismic_elements(seismic_elements);
-
-                // decrypting output
-                return self
-                    .inner
-                    .seismic_call(SendableTx::Builder(builder.clone()))
-                    .await
-                    .and_then(|encrypted_output| {
-                        // Decrypt the output using the encryption keypair
-                        let decrypted_output = seismic_elements
-                            .client_decrypt(
-                                &encrypted_output,
-                                &network_pk,
-                                &encryption_keypair.secret_key(),
-                            )
-                            .map_err(|e| {
-                                TransportErrorKind::custom_str(&format!(
-                                    "Error decrypting output: {:?}",
-                                    e
-                                ))
-                            })?;
-                        Ok(Bytes::from(decrypted_output))
-                    });
-            }
-        }
-        let res = self.inner.seismic_call(tx).await;
-        res
-    }
-
-    async fn send_transaction_internal(
-        &self,
-        mut tx: SendableTx<N>,
-    ) -> TransportResult<PendingTransactionBuilder<T, N>> {
-        if let Some(builder) = tx.as_mut_builder() {
-            if self.should_encrypt_input(builder) {
-                let network_pk = self.inner.get_tee_pubkey().await.map_err(|e| {
-                    TransportErrorKind::custom_str(&format!(
-                        "Error getting tee pubkey from server: {:?}",
-                        e
-                    ))
-                })?;
-                let encryption_keypair = TxSeismicElements::get_rand_encryption_keypair();
-                let seismic_elements = TxSeismicElements::default()
-                    .with_encryption_pubkey(encryption_keypair.public_key())
-                    .with_encryption_nonce(TxSeismicElements::get_rand_encryption_nonce());
-
-                // Encrypt using recipient's public key and generated private key
-                let plaintext_input = builder.input().unwrap();
-                let encrypted_input = seismic_elements
-                    .client_encrypt(&plaintext_input, &network_pk, &encryption_keypair.secret_key())
-                    .map_err(|e| {
-                        TransportErrorKind::custom_str(&format!("Error encrypting input: {:?}", e))
-                    })?;
-
-                builder.set_input(Bytes::from(encrypted_input));
-                builder.set_seismic_elements(seismic_elements);
-            }
-        }
-        let res = self.inner.send_transaction_internal(tx).await;
-        res
-    }
-}
-
+#[cfg(test)]
 /// Utilities for testing seismic provider
 pub mod test_utils {
     use alloy_primitives::{hex, Bytes};
@@ -350,14 +168,16 @@ pub mod test_utils {
 
 #[cfg(test)]
 mod tests {
-    use alloy_network::{Ethereum, EthereumWallet};
+    use alloy_network::{EthereumWallet, TransactionBuilder};
     use alloy_node_bindings::{Anvil, AnvilInstance};
-    use alloy_primitives::{Address, TxKind};
+    use alloy_primitives::{Address, Bytes, TxKind};
+    use alloy_rpc_types_eth::{TransactionInput, TransactionRequest};
     use alloy_signer_local::PrivateKeySigner;
 
-    use crate::test_utils::*;
-
-    use super::*;
+    use crate::{
+        layers::seismic::http::test_utils::ContractTestContext, Provider, SeismicSignedProvider,
+        SeismicUnsignedProvider, SendableTx,
+    };
 
     #[tokio::test]
     async fn test_seismic_signed_call() {
@@ -445,13 +265,5 @@ mod tests {
         let bob: PrivateKeySigner = anvil.keys()[1].clone().into();
         let wallet = EthereumWallet::from(bob.clone());
         wallet
-    }
-
-    #[tokio::test]
-    async fn test_get_tee_pubkey() {
-        let provider =
-            ProviderBuilder::new().network::<Ethereum>().layer(SeismicLayer {}).on_anvil();
-        let tee_pubkey = provider.get_tee_pubkey().await.unwrap();
-        println!("test_get_tee_pubkey: tee_pubkey: {:?}", tee_pubkey);
     }
 }
